💄 Consolidate game state management by introducing getStorage2048 and removing localStorage utility functions
